+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.3
 - Fix: Increment pay index count for non zaps [(@denis2342)](https://github.com/denis2342) 
 - Improvement: Add option to config for path of last pay index tip
+- Refactor: use filter map on relays to avoid second loop
 
 ## 0.1.2 
 - Improvement: Add preimage tag to zap note